+++ CHANGES.rst
@@ -8,6 +8,10 @@ Development version
   for instance with Pyodide
   https://github.com/joblib/joblib/pull/1256
 
+- Avoid unnecessary warnings when workers and main process delete
+  the temporary memmap folder contents concurrently.
+  https://github.com/joblib/joblib/pull/1263
+
 Release 1.1.0
 --------------